# contracts/v1/openapi.yaml
openapi: 3.1.0
info:
  title: TTSync MVP HTTP and SSE API
  version: 1.0.0
  description: MVP v1 的公开合同。房间命令使用 JSON expectedRevision，因此过期版本映射为 409 而不是 If-Match/412。SSE 仅作失效提示，客户端始终以角色化完整快照收敛。
servers:
  - url: /api
...
paths:
  /v1/accounts:
    post:
      security:
        - anonymousCsrf: []
      responses:
        "200":
          description: registerAccount 专属成功结果。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterAccountResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
      tags:
        - identity
      operationId: registerAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterAccountRequest"
      summary: 注册账号并发送验证邮件
  /v1/accounts/verification/resend:
    post:
      security:
        - anonymousCsrf: []
      responses:
        "200":
          description: resendVerification 专属成功结果。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResendVerificationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
      tags:
        - identity
      operationId: resendVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailRequest"
      summary: 以不暴露账号存在性的通用结果重发验证邮件
  /v1/accounts/verification:
    post:
      security:
        - anonymousCsrf: []
      responses:
        "200":
          description: verifyEmail 专属成功结果。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyEmailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
      tags:
        - identity
      operationId: verifyEmail
      description: 令牌无效、过期、重放、被新令牌取代或用途错误均返回相同的 422 VALIDATION_FAILED，不暴露具体原因。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRequest"
      summary: 使用一次性令牌验证邮箱
...
components:
  securitySchemes:
    anonymousCsrf:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: 未登录网页从同源 Cookie 取得令牌并通过 X-CSRF-Token 请求头回显。
    accountSession:
      type: apiKey
      in: cookie
      name: __Host-ttsync-session
      description: 登录会话；不缓存团队或房间角色。
    participantSession:
      type: apiKey
      in: cookie
      name: __Host-ttsync-participant
    spectatorSession:
      type: apiKey
      in: cookie
      name: __Host-ttsync-spectator
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
...
  schemas:
    Email:
      type: string
      format: email
      maxLength: 254
    Password:
      type: string
      minLength: 15
      maxLength: 128
      description: 15–128 个可打印 Unicode code point，不裁剪或静默截断。
...
    EmailRequest:
      type: object
      additionalProperties: false
      required:
        - email
      properties:
        email:
          $ref: "#/components/schemas/Email"
    TokenRequest:
      type: object
      additionalProperties: false
      required:
        - token
      properties:
        token:
          type: string
          minLength: 32
          maxLength: 512
    RegisterAccountRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          $ref: "#/components/schemas/Email"
        password:
          $ref: "#/components/schemas/Password"
        invitationToken:
          type: string
          minLength: 32
          maxLength: 512
...
    Violation:
      type: object
      additionalProperties: false
      required:
        - path
        - code
        - message
      properties:
        path:
          type: string
          description: 指向无效输入的 JSON Pointer。
          pattern: ^/
        code:
          type: string
          pattern: ^[A-Z][A-Z0-9_]*$
        message:
          type: string
          minLength: 1
...
    UnauthorizedProblem:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - detail
        - instance
        - requestId
        - status
        - code
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          minLength: 1
        detail:
          type: string
          minLength: 1
        instance:
          type: string
          format: uri-reference
        requestId:
          type: string
          format: uuid
        status:
          type: integer
          const: 401
        code:
          type: string
          const: UNAUTHORIZED
    ForbiddenProblem:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - detail
        - instance
        - requestId
        - status
        - code
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          minLength: 1
        detail:
          type: string
          minLength: 1
        instance:
          type: string
          format: uri-reference
        requestId:
          type: string
          format: uuid
        status:
          type: integer
          const: 403
        code:
          type: string
          const: FORBIDDEN
    ValidationProblem:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - detail
        - instance
        - requestId
        - status
        - code
        - violations
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          minLength: 1
        detail:
          type: string
          minLength: 1
        instance:
          type: string
          format: uri-reference
        requestId:
          type: string
          format: uuid
        status:
          type: integer
          const: 422
        code:
          type: string
          const: VALIDATION_FAILED
        violations:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Violation"
    InternalErrorProblem:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - detail
        - instance
        - requestId
        - status
        - code
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          minLength: 1
        detail:
          type: string
          minLength: 1
        instance:
          type: string
          format: uri-reference
        requestId:
          type: string
          format: uuid
        status:
          type: integer
          const: 500
        code:
          type: string
          const: INTERNAL_ERROR
    ServiceUnavailableProblem:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - detail
        - instance
        - requestId
        - status
        - code
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          minLength: 1
        detail:
          type: string
          minLength: 1
        instance:
          type: string
          format: uri-reference
        requestId:
          type: string
          format: uuid
        status:
          type: integer
          const: 503
        code:
          type: string
          const: SERVICE_UNAVAILABLE
    RateLimitedProblem:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - detail
        - instance
        - requestId
        - status
        - code
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          minLength: 1
        detail:
          type: string
          minLength: 1
        instance:
          type: string
          format: uri-reference
        requestId:
          type: string
          format: uuid
        status:
          type: integer
          const: 429
        code:
          type: string
          const: RATE_LIMITED
...
    RegisterAccountResponse:
      type: object
      additionalProperties: false
      required:
        - accepted
      properties:
        accepted:
          type: boolean
          const: true
    ResendVerificationResponse:
      type: object
      additionalProperties: false
      required:
        - accepted
      properties:
        accepted:
          type: boolean
          const: true
    VerifyEmailResponse:
      type: object
      additionalProperties: false
      required:
        - verified
      properties:
        verified:
          type: boolean
...
    MalformedRequestProblem:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - detail
        - instance
        - requestId
        - status
        - code
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          minLength: 1
        detail:
          type: string
          minLength: 1
        instance:
          type: string
          format: uri-reference
        requestId:
          type: string
          format: uuid
        status:
          type: integer
          const: 400
        code:
          type: string
          const: MALFORMED_REQUEST
    ResourceNotFoundProblem:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - detail
        - instance
        - requestId
        - status
        - code
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          minLength: 1
        detail:
          type: string
          minLength: 1
        instance:
          type: string
          format: uri-reference
        requestId:
          type: string
          format: uuid
        status:
          type: integer
          const: 404
        code:
          type: string
          const: RESOURCE_NOT_FOUND
    RoomNotFoundProblem:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - detail
        - instance
        - requestId
        - status
        - code
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          minLength: 1
        detail:
          type: string
          minLength: 1
        instance:
          type: string
          format: uri-reference
        requestId:
          type: string
          format: uuid
        status:
          type: integer
          const: 404
        code:
          type: string
          const: ROOM_NOT_FOUND
    RevisionConflictProblem:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - detail
        - instance
        - requestId
        - status
        - code
        - currentRevision
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          minLength: 1
        detail:
          type: string
          minLength: 1
        instance:
          type: string
          format: uri-reference
        requestId:
          type: string
          format: uuid
        status:
          type: integer
          const: 409
        code:
          type: string
          const: REVISION_CONFLICT
        currentRevision:
          type: integer
          minimum: 0
    IdempotencyConflictProblem:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - detail
        - instance
        - requestId
        - status
        - code
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          minLength: 1
        detail:
          type: string
          minLength: 1
        instance:
          type: string
          format: uri-reference
        requestId:
          type: string
          format: uuid
        status:
          type: integer
          const: 409
        code:
          type: string
          const: IDEMPOTENCY_CONFLICT
    StateConflictProblem:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - detail
        - instance
        - requestId
        - status
        - code
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          minLength: 1
        detail:
          type: string
          minLength: 1
        instance:
          type: string
          format: uri-reference
        requestId:
          type: string
          format: uuid
        status:
          type: integer
          const: 409
        code:
          type: string
          const: STATE_CONFLICT
  responses:
    BadRequest:
      description: 请求语法、JSON 或媒体类型无法解析（400）
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/MalformedRequestProblem"
          example:
            type: https://ttsync.example/problems/malformed-request
            title: 请求无法解析
            status: 400
            detail: JSON 请求体不完整。
            instance: /problems/req-400
            code: MALFORMED_REQUEST
            requestId: 018f5f60-c9c7-77e3-a6bb-08de53542952
    Unauthorized:
      description: 访问凭据缺失、无效或过期（401）
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/UnauthorizedProblem"
          example:
            type: https://ttsync.example/problems/unauthorized
            title: 需要有效凭据
            status: 401
            detail: 房间访问凭据缺失或已过期。
            instance: /problems/req-401
            code: UNAUTHORIZED
            requestId: 018f5f65-7191-71fd-b8a6-15c0c019d690
    Forbidden:
      description: 有效身份缺少当前操作权限（403）
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ForbiddenProblem"
          example:
            type: https://ttsync.example/problems/forbidden
            title: 没有操作权限
            status: 403
            detail: 当前角色没有此 capability。
            instance: /problems/req-403
            code: FORBIDDEN
            requestId: 018f5f6a-38c4-74a2-b0bb-eaf01a5d8f73
    NotFound:
      description: 资源不存在或对调用者不可见（404）
      content:
        application/problem+json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/ResourceNotFoundProblem"
              - $ref: "#/components/schemas/RoomNotFoundProblem"
            discriminator:
              propertyName: code
          example:
            type: https://ttsync.example/problems/not-found
            title: 房间不可见
            status: 404
            detail: 房间不存在或当前凭据不能查看。
            instance: /problems/req-404
            code: ROOM_NOT_FOUND
            requestId: 018f5f6f-80ae-7df8-9d5d-12277f280c57
    Conflict:
      description: 版本、幂等指纹/操作者或资源状态冲突（409）
      content:
        application/problem+json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/RevisionConflictProblem"
              - $ref: "#/components/schemas/IdempotencyConflictProblem"
              - $ref: "#/components/schemas/StateConflictProblem"
            discriminator:
              propertyName: code
          examples:
            revisionConflict:
              value:
                type: https://ttsync.example/problems/revision-conflict
                title: 房间已更新
                status: 409
                detail: 操作未应用，房间已更新。
                instance: /problems/req-1
                code: REVISION_CONFLICT
                requestId: 018f5f74-ddcb-76ef-8164-dfd82e2d96d8
                currentRevision: 9
            idempotencyConflict:
              value:
                type: https://ttsync.example/problems/idempotency-conflict
                title: 命令标识冲突
                status: 409
                detail: commandId 已由不同操作者或请求指纹使用。
                instance: /problems/req-2
                code: IDEMPOTENCY_CONFLICT
                requestId: 018f5f79-c84d-78f9-a50d-64b408158267
    UnprocessableEntity:
      description: 语法可解析但输入不满足合同或领域规则（422）
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ValidationProblem"
          example:
            type: https://ttsync.example/problems/validation-failed
            title: 输入校验失败
            status: 422
            detail: 两个字段无效。
            instance: /problems/req-3
            code: VALIDATION_FAILED
            requestId: 018f5f7f-b80c-711a-965b-ce09e4e75ba8
            violations:
              - path: /commandId
                code: INVALID_UUID
                message: 必须是 UUID。
              - path: /expectedRevision
                code: OUT_OF_RANGE
                message: 必须大于或等于 0。
    TooManyRequests:
      description: 请求超过限速（429）；调用者应遵循 Retry-After
      headers:
        Retry-After:
          required: true
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/RateLimitedProblem"
          example:
            type: https://ttsync.example/problems/rate-limited
            title: 请求过于频繁
            status: 429
            detail: 请在 Retry-After 指定时间后重试。
            instance: /problems/req-429
            code: RATE_LIMITED
            requestId: 018f5f84-e090-799c-8cf4-d7a22a0bfa67
    InternalServerError:
      description: 未预期的服务端失败（500）
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/InternalErrorProblem"
          example:
            type: https://ttsync.example/problems/internal-error
            title: 服务暂时失败
            status: 500
            detail: 请求未能完成，请使用 requestId 联系维护者。
            instance: /problems/req-500
            code: INTERNAL_ERROR
            requestId: 018f5f89-d59a-7e09-97c2-ca37b778be7d
    ServiceUnavailable:
      description: 邮件等必需依赖暂时不可用（503）；调用者可稍后重试
      headers:
        Retry-After:
          required: true
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ServiceUnavailableProblem"
          example:
            type: https://ttsync.example/problems/service-unavailable
            title: 服务暂时不可用
            status: 503
            detail: 邮件服务暂时不可用，请稍后重试。
            instance: /problems/req-503
            code: SERVICE_UNAVAILABLE
            requestId: 018f5f89-d59a-7e09-97c2-ca37b778be7e